requirements.txt
flask==3.0.0
pandas==2.1.4
numpy==1.26.2
plotly==5.18.0
dash==2.14.2
dash-bootstrap-components==1.5.0
requests==2.31.0
python-dotenv==1.0.0
sqlalchemy==2.0.23
scikit-learn==1.3.2
folium==0.15.1
geopy==2.4.1
schedule==1.2.0
gunicorn==21.2.0
pytest==7.4.3